Pointer size is usize not u32

+++ hw/model/src/xi3c/xi3c.rs
@@ -166,7 +166,7 @@ impl Controller {
         if status as i64 != 0 {}
     }
 
-    pub fn cfg_initialize(&mut self, config: &Config, effective_addr: u32) -> i32 {
+    pub fn cfg_initialize(&mut self, config: &Config, effective_addr: usize) -> i32 {
         if self.ready {
             return 5;
         }
